+++ Realm/RLMArrayTableView.mm
@@ -35,11 +35,10 @@
 
 + (instancetype)arrayWithObjectClassName:(NSString *)objectClassName
                                    view:(tightdb::TableView const &)view
-                                  realm:(RLMRealm *)realm{
-    RLMArrayTableView *ar = [[RLMArrayTableView alloc] initWithObjectClassName:objectClassName];
+                                  realm:(RLMRealm *)realm {
+    RLMArrayTableView *ar = [[RLMArrayTableView alloc] initViewWithObjectClassName:objectClassName readOnly:YES];
     ar->_backingView = view;
     ar->_realm = realm;
-    ar->_readOnly = YES;
     return ar;
 }