+++ Realm/RLMArray_Private.hpp
@@ -41,16 +41,19 @@
 }
 
 /**
- Initialize an RLMArray.
+ Initialize a standalone RLMArray.
  
- @warning Realm arrays are typed. You must specify an RLMObject class name during initialization
- and can only add objects of this type to the array.
+ @warning Realm arrays are typed. You must specify an RLMObject class name
+ during initialization and can only add objects of this type to the array.
  
  @param objectClassName     The class name of the RLMObjects this RLMArray will hold.
  
  @return                    An initialized RLMArray instance.
  */
 - (instancetype)initWithObjectClassName:(NSString *)objectClassName;
+
+// designated initializer for RLMArray subclasses
+- (instancetype)initViewWithObjectClassName:(NSString *)objectClassName readOnly:(BOOL)readOnly;
 
 // create standalone array variant
 + (instancetype)standaloneArrayWithObjectClassName:(NSString *)objectClassName;